Address review: docstring, errored-log location, test conventions
- Update _run_task docstring to describe the errored-EvalLog conversion
  (it no longer re-raises except under debug_errors)
- Set location on the synthesized errored EvalLog so consumers (view
  notification, telemetry, eval_retry) can reference the destination path
- Move the log-write-failure tests into tests/test_eval.py, add type
  annotations, and reword the log message to cover start-of-task writes
- Group the CHANGELOG entry with the other Bugfix items
